Match C++/Swift for Chapter 7

+++ raytrace/cpp/sphere.h
@@ -20,14 +20,14 @@ bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& rec) const 
     float c = dot(oc, oc) - radius*radius;
     float discriminant = b*b - a*c;
     if (discriminant > 0) {
-        float temp = (-b - sqrt(b*b-a*c))/a;
+        float temp = (-b - sqrtf(b*b-a*c))/a;
         if (temp < t_max && temp > t_min) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
             rec.normal = (rec.p - center) / radius;
             return true;
         }
-        temp = (-b + sqrt(b*b-a*c))/a;
+        temp = (-b + sqrtf(b*b-a*c))/a;
         if (temp < t_max && temp > t_min) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
